Move trie construction out to its own module
