+++ cranelift/isle/isle/src/lib.rs
@@ -26,3 +26,4 @@ pub mod ir;
 pub mod lexer;
 pub mod parser;
 pub mod sema;
+pub mod trie;
